// src/types/payment.ts
export interface CheckoutFormData {
  cardName: string;
  cardNumber: string;
  expiryMonth: string;
  expiryYear: string;
  cvc: string;
  saveCard?: boolean;
}

export interface PaymentIntent {
  order_id: string;
  amount: number;
  currency: string;
  description: string;
  payment_method: PaymentMethodType;
}

export interface PaymentSummary {
  subtotal: number;
  tax: number;
  discount: number;
  total: number;
  currency: string;
}